HW02: implemented login imitation

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -9,13 +9,14 @@ import ru.javawebinar.topjava.util.DateTimeUtil;
 import ru.javawebinar.topjava.util.MealsUtil;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
+
+import static ru.javawebinar.topjava.util.ValidationUtil.checkAuthorisation;
 
 @Repository
 public class InMemoryMealRepository implements MealRepository {
@@ -26,12 +27,13 @@ public class InMemoryMealRepository implements MealRepository {
     private InMemoryMealRepository() {
         MealsUtil.meals.forEach(meal -> {
             meal.setUserId(1);
-            save(meal);
+            save(meal, 1);
         });
     }
 
     @Override
-    public Meal save(Meal meal) {
+    public Meal save(Meal meal, int userId) {
+        checkAuthorisation(meal, userId);
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
             repository.put(meal.getId(), meal);
@@ -43,15 +45,16 @@ public class InMemoryMealRepository implements MealRepository {
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(int id, int userId) {
+        checkAuthorisation(get(id, userId), userId);
         log.info("delete {}", id);
         return repository.remove(id) != null;
     }
 
     @Override
-    public Meal get(int id) {
+    public Meal get(int id, int userId) {
         log.info("get {}", id);
-        return repository.get(id);
+        return checkAuthorisation(repository.getOrDefault(id, null), userId);
     }
 
     @Override
@@ -64,14 +67,11 @@ public class InMemoryMealRepository implements MealRepository {
     }
 
     @Override
-    public List<Meal> getAllFiltered(int userId, LocalDate startDate, LocalDate endDate, LocalTime startTime, LocalTime endTime) {
-        log.info("getAll for user {} from {} {} till {} {}", userId, startDate, startTime, endDate, endTime);
+    public List<Meal> getAllFilteredByDate(int userId, LocalDate startDate, LocalDate endDate) {
+        log.info("getAll for user {} from {} till {}", userId, startDate, endDate);
         return getAll(userId).stream()
-                .filter(meal -> DateTimeUtil.isBetweenDateOpen(meal.getDate(), startDate, endDate))
-                .filter(meal -> DateTimeUtil.isBetweenTimeHalfOpen(meal.getTime(), startTime, endTime))
+                .filter(meal -> DateTimeUtil.isBetween(meal.getDate(), startDate, endDate, true))
                 .collect(Collectors.toList());
     }
-
-
 }
